added text box to upload a image url for make post
