+++ client/src/components/AddComment.jsx
@@ -1,0 +1,21 @@
+import React from "react"
+
+const CommentText = () => {
+  return(
+
+    <div className="">
+        <label className="block text-sm font-medium leading-6 text-black">Make a comment!</label>
+        <div className="mt-1">
+            <textarea
+              id="comment"
+              name="comment"
+              rows={3}
+              className="mt-1 block w-full rounded-md border-0 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:py-1.5 sm:text-sm sm:leading-6"
+              placeholder="Enter your comment"
+              defaultValue={''}
+            />
+        </div>
+      </div>
+)
+}
+export default CommentText
