+++ client/src/components/Feed.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react'
 import { GetPosts } from '../services/PostServices'
-import { useNavigate, NavLink } from 'react-router-dom'
+import { useNavigate } from 'react-router-dom'
 // import { AiOutlineHeart, AiFillHeart } from 'react-icons/ai'
 import { Link } from 'react-router-dom'
-import MakeComment from './MakeComment'
+// import MakeComment from './MakeComment'
 
 function Feed({ user }) {
