+++ client/src/components/MakeComment.jsx
@@ -7,6 +7,7 @@ import { CreateComment } from '../services/PostServices'
 const MakeComment = ({ user }) => {
 let {postId} =useParams()
 const userName = localStorage.getItem('userName')
+console.log(userName)
 const userId = localStorage.getItem('userId')
 console.log(userId)
 
@@ -53,7 +54,7 @@ console.log(userId)
           <div className="grid gap-6">
             <div className="md:col-span-1">
               <h3 className="text-lg font-medium leading-6 text-gray-900">
-                Hey {formValues.userName},
+                Hey {formValues.userId},
               </h3>
               <p className="mt-1 text-sm text-gray-500">
                 Would you like to make a comment on this post?
