+++ client/src/components/ViewComments.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react'
 import { AiOutlineArrowLeft } from 'react-icons/ai'
 import { BsFillTrashFill } from 'react-icons/bs'
 import { Link, Navigate, useParams } from 'react-router-dom'
-
 import Client from '../services/api'
+import AddComment from './AddComment'
 
 const ViewComments = () => {
 
@@ -88,7 +88,7 @@ const ViewComments = () => {
 
 
 
-
+        <AddComment />
         </div>
